rag-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.3
httpx==0.27.0
SQLAlchemy[asyncio]==2.0.32
asyncpg==0.29.0
python-dotenv==1.0.1
sqladmin==0.21.0
wtforms==3.1.2
pgvector